backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
mysql-connector-python==8.3.0
pydantic==2.7.0
python-dotenv==1.0.1